+++ src/extension.ts
@@ -1,20 +1,21 @@
 import * as vscode from "vscode";
-import { prop } from "./utils";
-import { OpenBuffers } from "./open-buffers";
+import * as FindAlternate from "./FindAlternate";
+import { patterns } from "./patterns";
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
-  console.log('"vscode-open-buffers" is now active');
-
-  const buffers = vscode.window.visibleTextEditors.map(prop("document"));
-  const openBuffers = new OpenBuffers(buffers);
-
-  vscode.workspace.onDidOpenTextDocument(openBuffers.onBufferOpen);
-  vscode.workspace.onDidCloseTextDocument(openBuffers.onBufferClose);
+  console.log('"vscode-switch-to-spec" is now active');
 
   const commands = [
-    vscode.commands.registerCommand("open-buffers.openBuffer", openBuffers.openBuffer)
+    vscode.commands.registerCommand(
+      "switch-to-spec.alternateFile",
+      FindAlternate.openFile(patterns, { split: false })
+    ),
+    vscode.commands.registerCommand(
+      "switch-to-spec.alternateFileInSplit",
+      FindAlternate.openFile(patterns, { split: true })
+    )
   ];
 
   commands.forEach(command => context.subscriptions.push(command));