+++ src/test/utils.test.ts
@@ -7,4 +7,11 @@ suite("utils", () => {
     const array = [1, 2];
     assert.deepEqual(utils.cons(3, array), [3, 1, 2]);
   });
+
+  test("find", () => {
+    const greaterThan1 = (x: number) => x > 1;
+
+    assert.equal(utils.find(greaterThan1, [1, 2, 3]), 2);
+    assert.equal(utils.find(greaterThan1, [-1, -2]), null);
+  });
 });